jumpdest cache mechanism
