+++ tests/src/kakarot/test_account.py
@@ -1,7 +1,5 @@
 import pytest
-from ethereum.shanghai.vm.runtime import get_valid_jump_destinations
 
-from kakarot_scripts.utils.kakarot import get_contract
 from tests.utils.syscall_handler import SyscallHandler
 from tests.utils.uint256 import int_to_uint256
 
@@ -95,42 +93,3 @@ class TestAccount:
         ):
             output = cairo_run("test__has_code_or_nonce", nonce=nonce, code=code)
             assert output == expected_result
-
-    class TestGetJumpdests:
-        def test_should_return_same_as_execution_specs(self, cairo_run):
-            bytecode = get_contract("PlainOpcodes", "Counter").bytecode_runtime
-            output = cairo_run("test__get_jumpdests", bytecode=bytecode)
-            assert set(output) == get_valid_jump_destinations(bytecode)
-
-    class TestIsJumpdestValid:
-        def test_should_return_cached_valid_jumpdest(self, cairo_run):
-            assert (
-                cairo_run(
-                    "test__is_jumpdest_valid",
-                    cached_jumpdests=[0x01, 0x10, 0x101],
-                    query=0x10,
-                )
-                == 1
-            )
-            assert (
-                cairo_run(
-                    "test__is_jumpdest_valid",
-                    cached_jumpdests=[0x01, 0x10, 0x101],
-                    query=0x101,
-                )
-                == 1
-            )
-
-        @SyscallHandler.patch(
-            "IAccount.is_jumpdest_valid",
-            lambda addr, data: [1 if data == [0x10] else 0],
-        )
-        def test_should_return_non_cached_valid_jumpdest(self, cairo_run):
-            assert (
-                cairo_run("test__is_jumpdest_valid", cached_jumpdests=[], query=0x10)
-                == 1
-            )
-            assert (
-                cairo_run("test__is_jumpdest_valid", cached_jumpdests=[], query=0x102)
-                == 0
-            )
